G.729B: Tally up the return value of decode()/encode() in a variable
Once G.729B support is added, the number of audio samples decoded on
output will no longer be linearly correlated to the number of bytes fed
on input (or vice-versa for encoding).  We will therefore need to tally
up that number in a variable, which will act as a return value, as well
as an index for the destination buffer.