G.729B: Explicitly use 80:10 ratio in expressions, instead of 8:1
Once G.729B support is added, these expressions are about to get a bit
more complicated, to account for the different size of SID frames.  By
switching to a 80:10 ratio right away, we can get a head start and
reduce the complexity of the last patch.  (Not to mention that this
makes things a bit more explicit as a bonus.)

For the record, there is a reason behind performing the division before
the multiplication, which will become apparent in the final patch.  Note
that the quotient is guaranteed to be integral, due to the assertions.