Bison: Use `-Dapi.prefix` instead of `-p` when building the SDP parser
Bison's `-p` option, aside from being deprecated, does not have any
effect on `union YYSTYPE`, which ends up being defined twice (in
`parser/parser.hxx` and `sdp/sdp_parser.hxx`).  And while we never
include both headers in the same translation unit (which would be a
sufficient precaution in C), this is still a violation of the C++ One
Definition Rule.

Since changing the behavior of that option would have broken backwards
compatibility, `api.prefix` was introduced as its successor in version
2.6.¹  It does rename `YYSTYPE` (which we never refer to ourselves), as
well as a few other types and macros (which we don't use either), so it
doesn't break anything as far as we're concerned, while fixing our
initial issue.

 ¹ https://lists.gnu.org/archive/html/help-bison/2012-07/msg00009.html

(Note that while `api.prefix` was introduced in Bison 2.6, its syntax
changed in 3.0, which is therefore the minimum version we require.)