Bison: Drop old copy of `sdp_parser.h` and recreate it at build time
All files previously generated by Flex/Bison were removed by commit
d2f58e9, except for `sdp_parser.h`, which was apparently forgotten
behind.  As we're about to change the Bison invocation creating
`sdp_parser.yxx` in the next commit (which would then cause the two
files to drift out of sync), we now correct this omission.

(Just to be safe, I compared our own copy of this file with one freshly
rebuilt with Bison 2.3, confirming that they were equivalent and that
nothing would be lost with this change.)