qt6: Use `CMAKE_CXX_STANDARD` instead of `-std=`
Compiling against Qt 6 requires C++17 or later, which is incompatible
with our use of `-std=c++11`.  By switching to `CMAKE_CXX_STANDARD`, we
can still maintain C++11 as a minimum, while benefiting from an
automatic bump to a later version if need be.

(For the record, our code builds fine as-is up to C++20, and currently
only exhibits a single easy-to-fix failure with C++23.)

Note that this requires CMake 3.1, which is not available out-of-the-box
on the 10-year-old Ubuntu 14.04 (Trusty Tahr).