qt6: Use `${QT_VERSION_MAJOR}`
Instead of hardcoding our use of Qt 5 in several locations, this allows
us to only do it once, and then refer to "whichever version we found"
thereafter.  This will help when porting to Qt 6.

(Unfortunately, this does not work with Akonadi, so we'll need to get
our hands dirty in the final commit.)

Source: https://doc.qt.io/qt-6/cmake-qt5-and-qt6-compatibility.html