qt6: Add a `WITH_QT6` build option
Note that `WITH_QT5` and `WITH_QT6` are *not* currently exclusive, and
enabling the latter will not automatically disable the former.  If both
options are enabled, Qt 6 will be used if found, falling back to Qt 5
otherwise.  (This may change in the future.  It was just simpler to do
at the moment.)